Extract store into a separate file.
